+++ src/gtt_activation_dialog.c
@@ -26,7 +26,7 @@
 
 #include <qof.h>
 
-#include "cur-proj.h"
+#include "gtt_current_project.h"
 #include "gtt_help_popup.h"
 #include "gtt_project.h"
 #include "proj-query.h"
