+++ src/gtt_current_project.h
