+++ src/gtt_export.c
@@ -26,7 +26,7 @@
 #include <libgnomevfs/gnome-vfs.h>
 
 #include "app.h"
-#include "ghtml.h"
+#include "gtt_ghtml.h"
 #include "gtt_project.h"
 
 /* Project data export */
