+++ src/gtt_file_io.c
@@ -29,7 +29,7 @@
 #include <string.h>
 
 #include "app.h"
-#include "cur-proj.h"
+#include "gtt_current_project.h"
 #include "gtt_err_throw.h"
 #include "gtt_gsettings_io.h"
 #include "gtt.h"
@@ -38,7 +38,7 @@
 #include "gtt_preferences.h"
 #include "gtt_project.h"
 #include "proj_p.h"
-#include "timer.h"
+#include "gtt_timer.h"
 #include "gtt_toolbar.h"
 
 #ifdef USE_GTT_DEBUG_FILE
