+++ src/gtt_gsettings_io.c
@@ -26,14 +26,14 @@
 #include <gnome.h>
 
 #include "app.h"
-#include "cur-proj.h"
+#include "gtt_current_project.h"
 #include "gtt_gsettings_gnomeui.h"
 #include "gtt_gsettings_io_p.h"
 #include "gtt.h"
 #include "gtt_menus.h"
 #include "plug-in.h"
 #include "gtt_preferences.h"
-#include "timer.h"
+#include "gtt_timer.h"
 #include "gtt_toolbar.h"
 
 /* XXX these should not be externs, they should be part of
