+++ src/gtt_idle_dialog.c
@@ -32,7 +32,7 @@
 
 #include <qof.h>
 
-#include "cur-proj.h"
+#include "gtt_current_project.h"
 #include "gtt_help_popup.h"
 #include "gtt_project.h"
 #include "gtt_util.h"
