+++ src/gtt_project_queries.c
@@ -20,7 +20,7 @@
 
 #include <glib.h>
 
-#include "cur-proj.h"
+#include "gtt_current_project.h"
 #include "gtt_project.h"
 #include "proj-query.h"
