+++ src/gtt_projects_tree.c
@@ -31,7 +31,7 @@
 #include <glib/gi18n.h>
 
 #include "projects-tree.h"
-#include "timer.h"
+#include "gtt_timer.h"
 
 #define GTT_PROJECTS_TREE_GET_PRIVATE(obj) (G_TYPE_INSTANCE_GET_PRIVATE ((obj), GTT_TYPE_PROJECTS_TREE, GttProjectsTreePrivate))
