+++ src/gtt_queries.h
