+++ src/gtt_xml_write.c
@@ -24,7 +24,7 @@
 #include <qof.h>
 #include <stdio.h>
 
-#include "cur-proj.h"
+#include "gtt_current_project.h"
 #include "gtt_err_throw.h"
 #include "gtt.h"
 #include "gtt_project.h"
