+++ src/main.c
@@ -41,7 +41,7 @@
 #include <qof.h>
 
 #include "app.h"
-#include "cur-proj.h"
+#include "gtt_current_project.h"
 #include "gtt_err_throw.h"
 #include "gtt_file_io.h"
 #include "gtt.h"
@@ -50,7 +50,7 @@
 #include "menucmd.h"
 #include "gtt_preferences.h"
 #include "gtt_project.h"
-#include "timer.h"
+#include "gtt_timer.h"
 #include "gtt_toolbar.h"
 #include "gtt_xml.h"
